requirements.txt
fastapi
uvicorn
langchain-core>=0.1.33
langchain-community>=0.0.38
langchain-text-splitters>=0.0.1
langchain-pinecone>=0.1.3
pinecone-client>=3.0.0
python-multipart
pypdf
docx2txt
langdetect
ollama
langchain-groq>=0.1.7
sentence-transformers>=3.0.0
pydantic>=2.0.0
numpy==1.24.3
scipy==1.11.1
python-dotenv>=1.0.0
